% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.matrix.glmnet.r
\name{summary.matrix.glmnet}
\alias{summary.matrix.glmnet}
\title{Sumarize matrix_glmnet fit}
\usage{
\method{summary}{matrix.glmnet}(object, quantile.probs = c(0.01, 0.025, 0.5,
  0.975, 0.99))
}
\arguments{
\item{object}{Object from class \code{matrix.glmnet} as produced by \link{matrix_glmnet}.}

\item{quantile.probs}{Quantiles to be calculated from bootstrap. Only used if the method
element in \code{object} is bootstrap.}
}
\description{
This function takes a model fit with \link{matrix_glmnet} and calculates a series of
statistics depending on the model fit method.
}
\examples{
data(Rhizo)
data(Rhizo.map)
Dat <- create_dataset(Rhizo[1:10,],Rhizo.map)

set.seed(743)
m1 <- matrix_glmnet(Dat = Dat, formula = ~ fraction,
                   nperm = 100, family = "poisson",
                   method = "permutation",
                   verbose = FALSE)
m1.sum <- summary(m1)
m1.sum$coefficients
}
\author{
Sur from Dangl lab
}

